requirements.txt
apple_ocr==1.0.8
chromadb==1.0.20
docling==2.48.0
easyocr==1.7.2
huggingface_hub==0.34.3
layoutparser==0.3.4
mysql_connector_repackaged==0.3.1
openai==1.102.0
pandas==2.3.2
pdf2image==1.17.0
Pillow==9.5.0
Pillow==11.3.0
pytesseract==0.3.13
torch==2.7.0